Refactor packet framing to support generic streams
- introduce packets::uci::Reader and packets::uci::Writer
  which wrap around dyn AsyncRead and dyn AsyncWrite respectively
- the methods Reader::read and Writer::write perform the logic
  previoulsy implemented in Conntection::read and Connection::write
